% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Componentlayouter.R
\name{Componentlayouter}
\alias{Componentlayouter}
\alias{ComponentlayouterAddin}
\title{Componentlayouter}
\usage{
ComponentlayouterAddin()
}
\value{
\code{Componentlayouter} returns the layout as xy coordinates.
}
\description{
\code{Componentlayouter} is an RStudio-Addin that facilitates layouting networks with several components.
}
\details{
To run the addin, highlight an igraph-object in your current script and select \code{Componentlayouter} from the Addins-menu within RStudio.
}
