% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Netbuilder.R
\name{Netbuilder}
\alias{Netbuilder}
\alias{NetbuilderAddin}
\title{Netbuilder}
\usage{
NetbuilderAddin()
}
\value{
\code{Netbuilder} returns the created network as igraph object.
}
\description{
\code{Netbuilder} is an RStudio-Addin that allows you to create small networks with point and click.
}
\details{
To run the addin, select \code{Netbuilder} from the Addins-menu within RStudio.
}
