% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Netreader.R
\name{Netreader}
\alias{Netreader}
\alias{NetreaderAddin}
\title{Netreader}
\usage{
NetreaderAddin()
}
\value{
\code{Netreader} returns the created network as igraph object.
}
\description{
\code{Netreader} is an RStudio-Addin that allows you to read network files. So far, only plaintext files are supported (e.g. csv,tsv,...).
}
\details{
To run the addin, select \code{Netreader} from the Addins-menu within RStudio.
}
